[cxx-interop] Allow virtual methods to be renamed with swift_name
