+++ test/Interop/Cxx/class/inheritance/Inputs/virtual-methods.h
@@ -6,6 +6,8 @@ inline void testFunctionCollected() {
 
 struct Base {
   virtual void foo() = 0;
+  virtual void virtualRename() const
+      __attribute__((swift_name("swiftVirtualRename()")));
 };
 
 struct Base2 { virtual int f() = 0; };
@@ -49,48 +51,3 @@ struct DerivedFromCallsPureMethod : CallsPureMethod {
 };
 
 struct DerivedFromDerivedFromCallsPureMethod : DerivedFromCallsPureMethod {};
-
-struct HasDestructor {
-  ~HasDestructor() {}
-};
-
-// MARK: Reference Types:
-
-#define IMMORTAL_FRT                                                           \
-  __attribute__((swift_attr("import_reference")))                              \
-  __attribute__((swift_attr("retain:immortal")))                               \
-  __attribute__((swift_attr("release:immortal")))
-
-struct IMMORTAL_FRT ImmortalBase {
-  int value = 0;
-
-  virtual int get42() const { return 42; }
-  virtual int getOverridden42() const { return 123; }
-  virtual int getIntValue() const { return value; }
-};
-
-struct IMMORTAL_FRT Immortal : public ImmortalBase {
-  static Immortal *_Nonnull create() { return new Immortal(); }
-
-  virtual int getOverridden42() const override { return 42; }
-  virtual void setIntValue(int newValue) { this->value = newValue; }
-};
-
-struct IMMORTAL_FRT DerivedFromImmortal : public Immortal {
-  static DerivedFromImmortal *_Nonnull create() { return new DerivedFromImmortal(); }
-};
-
-struct IMMORTAL_FRT Immortal2 {
-public:
-  virtual void virtualMethod(HasDestructor) = 0;
-};
-
-inline const ImmortalBase *_Nonnull castToImmortalBase(
-    const Immortal *_Nonnull immortal) {
-  return static_cast<const ImmortalBase *>(immortal);
-}
-
-inline const Immortal *_Nonnull castToImmortal(
-    const DerivedFromImmortal *_Nonnull immortal) {
-  return static_cast<const Immortal *>(immortal);
-}
