+++ test/Interop/Cxx/class/inheritance/virtual-methods-typechecker.swift
@@ -15,25 +15,7 @@ let _ = DerivedFromDerived2()
 
 VirtualNonAbstractBase().nonAbstractMethod()
 
-@available(SwiftStdlib 5.8, *)
-func takesImmortalBase(_ i: ImmortalBase) {
-  let _ = i.get42()
-  let _ = i.getOverridden42()
-  let _ = i.getIntValue()
-}
-
-@available(SwiftStdlib 5.8, *)
-func takesImmortal(_ i: Immortal) {
-  let _ = i.get42()
-  let _ = i.getOverridden42()
-  let _ = i.getIntValue()
-  i.setIntValue(1)
-}
-
-@available(SwiftStdlib 5.8, *)
-func takesDerivedFromImmortal(_ i: DerivedFromImmortal) {
-  let _ = i.get42()
-  let _ = i.getOverridden42()
-  let _ = i.getIntValue()
-  i.setIntValue(1)
+func callVirtualRenamedMethod(_ b: Base) {
+  b.virtualRename()  // expected-error {{value of type 'Base' has no member 'virtualRename'}}
+  b.swiftVirtualRename()
 }
