+++ test/Interop/Cxx/foreign-reference/Inputs/member-inheritance.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#define FRT                                       \
+#define IMMORTAL_FRT                              \
   __attribute__((swift_attr("import_reference"))) \
   __attribute__((swift_attr("retain:immortal")))  \
   __attribute__((swift_attr("release:immortal")))
@@ -28,7 +28,7 @@ public:
     }
 private:
     int x;
-} FRT;
+} IMMORTAL_FRT;
 
 class CopyTrackedDerivedClass: public CopyTrackedBaseClass {
 public:
@@ -37,7 +37,7 @@ public:
     int getDerivedX() const {
         return getX();
     }
-} FRT;
+} IMMORTAL_FRT;
 
 CopyTrackedDerivedClass *makeCopyTrackedDerivedClass(int x) {
     return new CopyTrackedDerivedClass(x);
@@ -50,12 +50,12 @@ public:
     }
 private:
     int y = 11;
-} FRT;
+} IMMORTAL_FRT;
 
 class CopyTrackedDerivedDerivedClass: public NonEmptyBase, public CopyTrackedDerivedClass {
 public:
     CopyTrackedDerivedDerivedClass(int x) : CopyTrackedDerivedClass(x) {}
-} FRT;
+} IMMORTAL_FRT;
 
 CopyTrackedDerivedDerivedClass *makeCopyTrackedDerivedDerivedClass(int x) {
     return new CopyTrackedDerivedDerivedClass(x);
@@ -66,15 +66,14 @@ public:
     CopyTrackedDerivedClass &operator[](int x) const {
         return *new CopyTrackedDerivedClass(x);
     }
-} FRT;
+} IMMORTAL_FRT;
 
 BaseReturningFRTFromSubscript *makeBaseReturningFRTFromSubscript() {
     return new BaseReturningFRTFromSubscript();
 }
 
-class DerivedFromBaseReturningFRTFromSubscript: public BaseReturningFRTFromSubscript {
-public:
-} FRT;
+class DerivedFromBaseReturningFRTFromSubscript
+    : public BaseReturningFRTFromSubscript{public : } IMMORTAL_FRT;
 
 DerivedFromBaseReturningFRTFromSubscript *makeDerivedFromBaseReturningFRTFromSubscript() {
     return new DerivedFromBaseReturningFRTFromSubscript();
@@ -90,12 +89,53 @@ public:
 
 private:
     CopyTrackedDerivedClass *value;
-} FRT;
+} IMMORTAL_FRT;
 
-class DerivedFromBaseReturningFRTFromSubscriptPointer: public BaseReturningFRTFromSubscriptPointer {
-public:
-} FRT;
+class DerivedFromBaseReturningFRTFromSubscriptPointer
+    : public BaseReturningFRTFromSubscriptPointer{public : } IMMORTAL_FRT;
 
 DerivedFromBaseReturningFRTFromSubscriptPointer *makeDerivedFromBaseReturningFRTFromSubscriptPointer() {
     return new DerivedFromBaseReturningFRTFromSubscriptPointer();
+}
+
+struct IMMORTAL_FRT ImmortalBase {
+  int value = 0;
+
+  virtual int get42() const { return 42; }
+  virtual int getOverridden42() const { return 123; }
+  virtual int getIntValue() const { return value; }
+};
+
+struct IMMORTAL_FRT Immortal : public ImmortalBase {
+  static Immortal *_Nonnull create() { return new Immortal(); }
+
+  virtual int getOverridden42() const override { return 42; }
+  virtual void setIntValue(int newValue) { this->value = newValue; }
+};
+
+struct IMMORTAL_FRT DerivedFromImmortal : public Immortal {
+  static DerivedFromImmortal *_Nonnull create() {
+    return new DerivedFromImmortal();
+  }
+};
+
+struct HasDestructor {
+  ~HasDestructor() {}
+};
+
+struct IMMORTAL_FRT Immortal2 {
+public:
+  virtual void virtualMethod(HasDestructor) = 0;
+  virtual void virtualRename() const
+      __attribute__((swift_name("swiftVirtualRename()")));
+};
+
+inline const ImmortalBase *_Nonnull castToImmortalBase(
+    const Immortal *_Nonnull immortal) {
+  return static_cast<const ImmortalBase *>(immortal);
+}
+
+inline const Immortal *_Nonnull castToImmortal(
+    const DerivedFromImmortal *_Nonnull immortal) {
+  return static_cast<const Immortal *>(immortal);
 }
