+++ test/Interop/Cxx/foreign-reference/member-inheritance-typechecker.swift
@@ -1,0 +1,34 @@
+// RUN: %target-typecheck-verify-swift -verify-ignore-unknown -I %S/Inputs -cxx-interoperability-mode=upcoming-swift
+// RUN: %target-typecheck-verify-swift -verify-ignore-unknown -I %S/Inputs -cxx-interoperability-mode=swift-5.9
+// RUN: %target-typecheck-verify-swift -verify-ignore-unknown -I %S/Inputs -cxx-interoperability-mode=swift-6
+
+import MemberInheritance
+
+@available(SwiftStdlib 5.8, *)
+func takesImmortalBase(_ i: ImmortalBase) {
+  let _ = i.get42()
+  let _ = i.getOverridden42()
+  let _ = i.getIntValue()
+}
+
+@available(SwiftStdlib 5.8, *)
+func takesImmortal(_ i: Immortal) {
+  let _ = i.get42()
+  let _ = i.getOverridden42()
+  let _ = i.getIntValue()
+  i.setIntValue(1)
+}
+
+@available(SwiftStdlib 5.8, *)
+func takesDerivedFromImmortal(_ i: DerivedFromImmortal) {
+  let _ = i.get42()
+  let _ = i.getOverridden42()
+  let _ = i.getIntValue()
+  i.setIntValue(1)
+}
+
+@available(SwiftStdlib 5.8, *)
+func callsRenamedVirtualMethodsInFRT(_ i: Immortal2) {
+  i.virtualRename()  // expected-error {{value of type 'Immortal2' has no member 'virtualRename'}}
+  i.swiftVirtualRename()
+}
